+++ src/Symfony/Component/Translation/Formatter/MessageFormatter.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Translation\Formatter;
 
 use Symfony\Component\Translation\IdentityTranslator;
 use Symfony\Component\Translation\MessageSelector;
+use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
@@ -29,7 +30,7 @@ class MessageFormatter implements MessageFormatterInterface, ChoiceMessageFormat
     {
         if ($translator instanceof MessageSelector) {
             $translator = new IdentityTranslator($translator);
-        } elseif (null !== $translator && !$translator instanceof TranslatorInterface) {
+        } elseif (null !== $translator && !$translator instanceof TranslatorInterface && !$translator instanceof LegacyTranslatorInterface) {
             throw new \TypeError(sprintf('Argument 1 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
         }
 
@@ -41,15 +42,27 @@ class MessageFormatter implements MessageFormatterInterface, ChoiceMessageFormat
      */
     public function format($message, $locale, array $parameters = array())
     {
+        if ($this->translator instanceof TranslatorInterface) {
+            return $this->translator->trans($message, $parameters, null, $locale);
+        }
+
         return strtr($message, $parameters);
     }
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated since Symfony 4.2, use format() with a %count% parameter instead
      */
     public function choiceFormat($message, $number, $locale, array $parameters = array())
     {
-        $parameters = array_merge(array('%count%' => $number), $parameters);
+        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.2, use the format() one instead with a %count% parameter.', __METHOD__), E_USER_DEPRECATED);
+
+        $parameters = array('%count%' => $number) + $parameters;
+
+        if ($this->translator instanceof TranslatorInterface) {
+            return $this->format($message, $locale, $parameters);
+        }
 
         return $this->format($this->translator->transChoice($message, $number, array(), null, $locale), $locale, $parameters);
     }