+++ src/Symfony/Component/Translation/IdentityTranslator.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\Translation;
 
+use Symfony\Component\Translation\TranslatorInterface as LegacyTranslatorInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 use Symfony\Contracts\Translation\TranslatorTrait;
 
 /**
@@ -18,11 +20,9 @@ use Symfony\Contracts\Translation\TranslatorTrait;
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
-class IdentityTranslator implements TranslatorInterface
+class IdentityTranslator implements LegacyTranslatorInterface, TranslatorInterface
 {
-    use TranslatorTrait {
-        transChoice as private doTransChoice;
-    }
+    use TranslatorTrait;
 
     private $selector;
 
@@ -40,14 +40,18 @@ class IdentityTranslator implements TranslatorInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated since Symfony 4.2, use the trans() method instead with a %count% parameter
      */
     public function transChoice($id, $number, array $parameters = array(), $domain = null, $locale = null)
     {
+        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.2, use the trans() one instead with a "%count%" parameter.', __METHOD__), E_USER_DEPRECATED);
+
         if ($this->selector) {
-            return strtr($this->selector->choose((string) $id, (int) $number, $locale ?: $this->getLocale()), $parameters);
+            return strtr($this->selector->choose((string) $id, $number, $locale ?: $this->getLocale()), $parameters);
         }
 
-        return $this->doTransChoice($id, $number, $parameters, $domain, $locale);
+        return $this->trans($id, array('%count%' => $number) + $parameters, $domain, $locale);
     }
 
     private function getPluralizationRule(int $number, string $locale): int