+++ src/Symfony/Component/Translation/LoggingTranslator.php
@@ -32,8 +32,11 @@ class LoggingTranslator implements LegacyTranslatorInterface, TranslatorBagInter
      * @param TranslatorInterface $translator The translator must implement TranslatorBagInterface
      * @param LoggerInterface     $logger
      */
-    public function __construct(TranslatorInterface $translator, LoggerInterface $logger)
+    public function __construct($translator, LoggerInterface $logger)
     {
+        if (!$translator instanceof LegacyTranslatorInterface && !$translator instanceof TranslatorInterface) {
+            throw new \TypeError(sprintf('Argument 1 passed to %s() must be an instance of %s, %s given.', __METHOD__, TranslatorInterface::class, \is_object($translator) ? \get_class($translator) : \gettype($translator)));
+        }
         if (!$translator instanceof TranslatorBagInterface) {
             throw new InvalidArgumentException(sprintf('The Translator "%s" must implement TranslatorInterface and TranslatorBagInterface.', \get_class($translator)));
         }
@@ -55,10 +58,19 @@ class LoggingTranslator implements LegacyTranslatorInterface, TranslatorBagInter
 
     /**
      * {@inheritdoc}
+     *
+     * @deprecated since Symfony 4.2, use the trans() method instead with a %count% parameter
      */
     public function transChoice($id, $number, array $parameters = array(), $domain = null, $locale = null)
     {
-        $trans = $this->translator->transChoice($id, $number, $parameters, $domain, $locale);
+        @trigger_error(sprintf('The "%s()" method is deprecated since Symfony 4.2, use the trans() one instead with a "%count%" parameter.', __METHOD__), E_USER_DEPRECATED);
+
+        if ($this->translator instanceof TranslatorInterface) {
+            $trans = $this->translator->trans($id, array('%count%' => $number) + $parameters, $domain, $locale);
+        } else {
+            $trans = $this->translator->transChoice($id, $number, $parameters, $domain, $locale);
+        }
+
         $this->log($id, $domain, $locale);
 
         return $trans;