+++ src/Symfony/Component/Translation/Tests/DataCollectorTranslatorTest.php
@@ -25,7 +25,7 @@ class DataCollectorTranslatorTest extends TestCase
 
         $collector->trans('foo');
         $collector->trans('bar');
-        $collector->transChoice('choice', 0);
+        $collector->trans('choice', array('%count%' => 0));
         $collector->trans('bar_ru');
         $collector->trans('bar_ru', array('foo' => 'bar'));
 
@@ -54,7 +54,7 @@ class DataCollectorTranslatorTest extends TestCase
             'locale' => 'en',
             'domain' => 'messages',
             'state' => DataCollectorTranslator::MESSAGE_MISSING,
-            'parameters' => array(),
+            'parameters' => array('%count%' => 0),
             'transChoiceNumber' => 0,
         );
         $expectedMessages[] = array(
@@ -74,6 +74,30 @@ class DataCollectorTranslatorTest extends TestCase
             'state' => DataCollectorTranslator::MESSAGE_EQUALS_FALLBACK,
             'parameters' => array('foo' => 'bar'),
             'transChoiceNumber' => null,
+        );
+
+        $this->assertEquals($expectedMessages, $collector->getCollectedMessages());
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testCollectMessagesTransChoice()
+    {
+        $collector = $this->createCollector();
+        $collector->setFallbackLocales(array('fr', 'ru'));
+        $collector->transChoice('choice', 0);
+
+        $expectedMessages = array();
+
+        $expectedMessages[] = array(
+              'id' => 'choice',
+              'translation' => 'choice',
+              'locale' => 'en',
+              'domain' => 'messages',
+              'state' => DataCollectorTranslator::MESSAGE_MISSING,
+              'parameters' => array('%count%' => 0),
+              'transChoiceNumber' => 0,
         );
 
         $this->assertEquals($expectedMessages, $collector->getCollectedMessages());