+++ src/Symfony/Component/Translation/TranslatorInterface.php
@@ -11,13 +11,59 @@
 
 namespace Symfony\Component\Translation;
 
-use Symfony\Contracts\Translation\TranslatorInterface as BaseTranslatorInterface;
+use Symfony\Component\Translation\Exception\InvalidArgumentException;
 
 /**
+ * TranslatorInterface.
+ *
  * @author Fabien Potencier <fabien@symfony.com>
  *
- * @deprecated since Symfony 4.2, use the same interface from the Symfony\Contracts\Translation namespace
+ * @deprecated since Symfony 4.2, use Symfony\Contracts\Translation\TranslatorInterface instead
  */
-interface TranslatorInterface extends BaseTranslatorInterface
+interface TranslatorInterface
 {
+    /**
+     * Translates the given message.
+     *
+     * @param string      $id         The message id (may also be an object that can be cast to string)
+     * @param array       $parameters An array of parameters for the message
+     * @param string|null $domain     The domain for the message or null to use the default
+     * @param string|null $locale     The locale or null to use the default
+     *
+     * @return string The translated string
+     *
+     * @throws InvalidArgumentException If the locale contains invalid characters
+     */
+    public function trans($id, array $parameters = array(), $domain = null, $locale = null);
+
+    /**
+     * Translates the given choice message by choosing a translation according to a number.
+     *
+     * @param string      $id         The message id (may also be an object that can be cast to string)
+     * @param int         $number     The number to use to find the indice of the message
+     * @param array       $parameters An array of parameters for the message
+     * @param string|null $domain     The domain for the message or null to use the default
+     * @param string|null $locale     The locale or null to use the default
+     *
+     * @return string The translated string
+     *
+     * @throws InvalidArgumentException If the locale contains invalid characters
+     */
+    public function transChoice($id, $number, array $parameters = array(), $domain = null, $locale = null);
+
+    /**
+     * Sets the current locale.
+     *
+     * @param string $locale The locale
+     *
+     * @throws InvalidArgumentException If the locale contains invalid characters
+     */
+    public function setLocale($locale);
+
+    /**
+     * Returns the current locale.
+     *
+     * @return string The locale
+     */
+    public function getLocale();
 }