+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
    the "APP_ENV" environment variable instead.
  * Deprecated the `--no-debug` console option, set the "APP_DEBUG" 
    environment variable to "0" instead.
+ * Deprecated the `Templating\Helper\TranslatorHelper::transChoice()` method, use the `trans()` one instead with a `%count%` parameter
 
 4.1.0
 -----